refactor(timetable): extract pure form helpers and field components into event-form/
Mechanical, behavior-preserving extraction from timetable-event-modal.tsx
(3594 -> 3008 lines) in preparation for the three-step wizard split:

- event-form/form-model.ts: RepeatMode, EventFormState, PersonOption,
  STUDENT_PAGE_SIZE, isoWeekday, emptyForm, formFromInstance,
  formFromSeries, parseRequiredStaffOverride, sortPeople,
  schoolClassLabel, targetCohortActionLabel, initialStudentIDs,
  initialStaffIDs, initialPrimaryStaffID, fetchAllStudentOptions
- event-form/field.tsx: local Field component (1:1, keeps its
  text-red-600 for now; cleanup happens in the design-debt chunk)
- event-form/multi-select-field.tsx: local MultiSelectField (1:1)

plan-design-guards.test.ts: listTsxFiles is now recursive so the new
event-form/ subdirectory stays under the guard; bright-color allowlist
split accordingly (timetable-event-modal.tsx 14 -> 13, event-form/field.tsx 1,
total unchanged).

timetable-event-modal.test.tsx is untouched; all 87 modal tests and
138 design-guard tests pass unchanged.
